+++ R/dd_ip.R
@@ -1,0 +1,6 @@
+#' Delay Discounting Data
+#'
+#' A dataset containing a set of fake delay discounting responses
+#'
+#' @format A data frame with delay discounting responses
+"dd_ip"