+++ data-raw/data.R
@@ -16,3 +16,20 @@ five.fivetrial_pd <- five.fivetrial_pd[five.fivetrial_pd$ResponseId %in%
 five.fivetrial_pd$ResponseId <- 1:4
 
 usethis::use_data(five.fivetrial_pd, overwrite = TRUE)
+
+## make mock discounting data
+set.seed(123)
+num_participants <- 100
+delay_values <- c(1, 7, 30, 90, 180, 365)
+
+dd_ip <- do.call(rbind, lapply(1:num_participants, function(participant) {
+  k <- runif(1, 0.001, 1) # random discount rate for each participant
+  y <- 1 / (1 + k * delay_values) + rnorm(length(delay_values), 0, 0.05) # Indifference points
+  data.frame(
+    id = paste0("P", participant),
+    x = delay_values,
+    y = pmax(0, y) # indifference points are non-negative
+  )
+}))
+
+usethis::use_data(dd_ip, overwrite = TRUE)